pre 5.0 release

+++ Public/Remove-LMCachedAccount.ps1
@@ -1,36 +1,39 @@
 Function Remove-LMCachedAccount {
-    [CmdletBinding()]
+    [CmdletBinding(SupportsShouldProcess,ConfirmImpact='High')]
     Param (
         [Parameter(Mandatory, ParameterSetName = 'Single', ValueFromPipelineByPropertyName)]
         [Alias("Portal")]
         [String]$CachedAccountName,
 
         [Parameter(ParameterSetName = 'All')]
-        [Boolean]$RemoveAllEntries = $false
+        [Switch]$RemoveAllEntries
     )
     Begin{}
     Process{
         If($RemoveAllEntries){
             $CachedAccounts = Get-SecretInfo -Vault Logic.Monitor
-    
-            Foreach ($Account in $CachedAccounts.Name){
+            if ($PSCmdlet.ShouldProcess("$(($CachedAccounts | Measure-Object).Count) cached account(s)", "Remove All Cached Accounts")) {                
+                Foreach ($Account in $CachedAccounts.Name){
+                    Try{
+                        Remove-Secret -Name $Account -Vault Logic.Monitor -Confirm:$false -ErrorAction Stop
+                        Write-Host "[INFO]: Removed cached account secret for: $Account"
+                    }
+                    Catch{
+                        Write-Error $_.Exception.Message
+                    }
+                }
+                Write-Host "[INFO]: Processed all entries from credential cache"
+            }
+        }
+        Else{
+            If ($PSCmdlet.ShouldProcess($CachedAccountName, "Remove Cached Account")) {                
                 Try{
-                    Remove-Secret -Name $Account -Vault Logic.Monitor -Confirm:$false -ErrorAction Stop
-                    Write-Host "Removed cached account secret for: $Account"
+                    Remove-Secret -Name $CachedAccountName -Vault Logic.Monitor -Confirm:$false -ErrorAction Stop
+                    Write-Host "[INFO]: Removed cached account secret for: $CachedAccountName"
                 }
                 Catch{
                     Write-Error $_.Exception.Message
                 }
-            }
-            Write-Host "Processed all entries from credential cache"
-        }
-        Else{
-            Try{
-                Remove-Secret -Name $CachedAccountName -Vault Logic.Monitor -Confirm:$false -ErrorAction Stop
-                Write-Host "Removed cached account secret for: $CachedAccountName"
-            }
-            Catch{
-                Write-Error $_.Exception.Message
             }
         }
     }
